pgload/pyproject.toml
[project]
name = "pgload"
version = "0.1.0"
description = ""
authors = [{ name = "raczu", email = "Raczuekk@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer (>=0.15.2,<0.16.0)",
    "faker (>=37.3.0,<38.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "prometheus-client (>=0.22.0,<0.23.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["pgload"]

[tool.ruff.lint]
select = ["B", "C", "E", "F", "I", "PIE", "W", "TID"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.10"
black = "^25.1.0"

[project.scripts]
pgload = "pgload.cli:app"
